Modificaciones varias

+++ car.js
@@ -14,6 +14,8 @@ document.addEventListener("DOMContentLoaded", function () {
         }
     }
 
+    // MOSTRAR PRODUCTOS AGREGADOS EN EL CARRITO//
+
     function mostrarProductos(productosPorCategoria) {
         seccion__carrito.innerHTML = "";
         productosPorCategoria.forEach(producto => {
@@ -49,6 +51,8 @@ document.addEventListener("DOMContentLoaded", function () {
         actualizartTotal();
     }
 
+    // ELIMINAR PRODUCTOS DEL CARRITO//
+
     function botonEliminar() {
         eliminarArticulos = document.querySelectorAll(".c__eliminar");
         eliminarArticulos.forEach(boton => {
@@ -68,19 +72,41 @@ document.addEventListener("DOMContentLoaded", function () {
 
     cargarProductos();
 
+    // VACIAR EL CARRITO//
+
     vaciarCarrito.addEventListener("click", vaciar);
     function vaciar() {
 
-        productosEnCarrito.length = 0;
-        localStorage.setItem("productosEnCarrito", JSON.stringify(productosEnCarrito));
-        cargarProductos();
-        actualizarEstadoCarritoVacio();
+    // USO DE LIBRERIA SWEETALERT AL VACIAR CARRITO//
+
+        Swal.fire({
+            text: "¿Está seguro de que desea vaciar el carrito?",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#3085d6',
+            cancelButtonText: 'Cancelar',
+            confirmButtonText: 'Confirmar'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                productosEnCarrito.length = 0;
+                localStorage.setItem("productosEnCarrito", JSON.stringify(productosEnCarrito));
+                cargarProductos();
+                actualizarEstadoCarritoVacio();
+            }
+        });
     }
+
+    // CALCULAR EL PRECIO TOTAL DE PRODUCTOS EN EL CARRITO//
+
     function actualizartTotal() {
         const totalSuma = productosEnCarrito.reduce((acumulador, producto) => acumulador + (producto.precio * producto.cantidad), 0);
         mostrarTotal.innerText = `$${totalSuma}`;
         actualizarCarro();
     }
+
+    // CANTIDAD TOTAL DE PRODUCTOS EN EL CARRITO//
+
     function actualizarCarro() {
         let numCar = productosEnCarrito.reduce((acumulador, producto) => acumulador + producto.cantidad, 0);
         numCarrito.innerText = numCar;
